fix solo.json generation

+++ chef-repo/deploy.sh
@@ -45,13 +45,19 @@ ${CLOUDOS_BASE}/cloudos-lib/chef-repo/cookbooks \
 $(find ${CLOUDOS_BASE}/cloudos-apps/apps -type d -name cookbooks) \
 "
 
-SOLO_JSON="$(pwd)/solo-base.json"
-TMP_SOLO=$(mktemp /tmp/solo-json.XXXXXX) || die "Error creating temp file for solo.json"
+SOLO_JSON="${BASE}/solo-base.json"
+TMP_SOLO=$(mktemp /tmp/cloudos-solo-json.XXXXXX) || die "Error creating temp file for solo.json"
+
+CLOUDOS_INIT_BAG="${INIT_FILES}/data_bags/cloudos/init.json"
+if [ ! -f ${CLOUDOS_INIT_BAG} ] ; then
+  die "No ${CLOUDOS_INIT_BAG} found"
+fi
 
 # If not using Dyn directly from cloudos, install cloudos-dns
 JSON_EDIT="java -cp ${BASE}/../target/cloudos-server-*.jar org.cobbzilla.util.json.main.JsonEditor"
-DYN_ZONE=$(${JSON_EDIT} -f data_bags/cloudos/init.json -o read -p dns.zone | tr -d ' ')
-if [ -z ${DYN_ZONE} ] ; then
+DYN_ZONE=$(${JSON_EDIT} -f ${CLOUDOS_INIT_BAG} -o read -p dns.zone | tr -d ' ')
+if [ -z "${DYN_ZONE}" ] ; then
+  echo "No DynDNS config detected, enabling cloudos-dns..."
   REQUIRED="${REQUIRED} \
     data_bags/cloudos-dns/init.json \
     data_bags/cloudos-dns/ports.json"
@@ -63,6 +69,7 @@ else
 fi
 
 # Add cloudos recipe
-${JSON_EDIT} -f ${TMP_SOLO} -o write -p run_list[] -v \"recipe[cloudos]\" > ${TMP_SOLO}
+SOLO_JSON=$(mktemp /tmp/cloudos-solo-json.XXXXXX) || die "Error creating final temp file for solo.json"
+${JSON_EDIT} -f ${TMP_SOLO} -o write -p run_list[] -v \"recipe[cloudos]\" > ${SOLO_JSON}
 
-${DEPLOYER} ${host} ${INIT_FILES} "${REQUIRED}" "${COOKBOOK_SOURCES}" ${TMP_SOLO}
+${DEPLOYER} ${host} ${INIT_FILES} "${REQUIRED}" "${COOKBOOK_SOURCES}" ${SOLO_JSON}
